submission ready

+++ utils/extract.py
@@ -86,7 +86,7 @@ class WebExtractor:
             data["Timestamp"] = datetime.datetime.now()
 
             # Convert to rupiah
-            data["Price"] = int(data["Price"]) * 16000
+            data["Price"] = float(data["Price"]) * 16000
 
         except Exception as e:
             print(f"Failed to extract data from card. Error: {e}")
